backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
cryptography==43.0.1
PyMuPDF==1.24.11
python-docx==1.1.2
beautifulsoup4==4.12.3
markdown==3.7
rank-bm25==0.2.2
jieba==0.42.1
openai==1.51.0
